requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
pymysql
sqlite-utils
oracledb
pyodbc
pytest
httpx
questionary
